requirements.txt
opencv-python
progress_bar
scenedetect
tensorflow
keras
pillow
torch===1.3.0+cpu
torchvision===0.2.2.post3
numpy
git+https://github.com/GantMan/nsfw_model.git